Separate enumeration tests from derivation mode tests
