+++ tests/src/test/scala/pureconfig/DerivationModesSuite.scala
@@ -1,12 +1,9 @@
 package pureconfig
 
-import com.typesafe.config.{ ConfigFactory, ConfigValueFactory, ConfigValueType }
-import pureconfig.DerivationModesSuite._
-import pureconfig.error.WrongType
-import pureconfig.generic.error.NoValidCoproductChoiceFound
+import com.typesafe.config.ConfigFactory
 import shapeless.test.illTyped
 
-object DerivationModesSuite {
+class DerivationModesSuite extends BaseSuite {
 
   sealed trait Entity
   case class Person(name: String, surname: String) extends Entity
@@ -14,13 +11,6 @@ object DerivationModesSuite {
 
   val person = Person("John", "Doe")
   val conf = ConfigFactory.parseString("{ type: person, name: John, surname: Doe }")
-
-  sealed trait Color
-  case object RainyBlue extends Color
-  case object SunnyYellow extends Color
-}
-
-class DerivationModesSuite extends BaseSuite {
 
   behavior of "default"
 
@@ -65,76 +55,6 @@ class DerivationModesSuite extends BaseSuite {
 
     ConfigReader[Entity].from(conf.root) shouldBe Right(person)
     ConfigWriter[Entity].to(person) shouldBe conf.root()
-  }
-
-  it should "provide methods to derive readers for enumerations encoded as sealed traits" in {
-    import pureconfig.generic.semiauto._
-
-    implicit val colorReader = deriveEnumerationReader[Color]
-
-    ConfigReader[Color].from(ConfigValueFactory.fromAnyRef("rainy-blue")) shouldBe Right(RainyBlue)
-    ConfigReader[Color].from(ConfigValueFactory.fromAnyRef("sunny-yellow")) shouldBe Right(SunnyYellow)
-
-    val unknownValue = ConfigValueFactory.fromAnyRef("blue")
-    ConfigReader[Color].from(unknownValue) should failWith(NoValidCoproductChoiceFound(unknownValue), "")
-    ConfigReader[Color].from(conf.root()) should failWith(WrongType(ConfigValueType.OBJECT, Set(ConfigValueType.STRING)), "")
-  }
-
-  it should "provide methods to derive writers for enumerations encoded as sealed traits" in {
-    import pureconfig.generic.semiauto._
-
-    implicit val colorWriter = deriveEnumerationWriter[Color]
-
-    ConfigWriter[Color].to(RainyBlue) shouldEqual ConfigValueFactory.fromAnyRef("rainy-blue")
-    ConfigWriter[Color].to(SunnyYellow) shouldEqual ConfigValueFactory.fromAnyRef("sunny-yellow")
-  }
-
-  it should "provide methods to derive full converters for enumerations encoded as sealed traits" in {
-    import pureconfig.generic.semiauto._
-
-    implicit val colorConvert = deriveEnumerationConvert[Color]
-
-    ConfigConvert[Color].from(ConfigValueFactory.fromAnyRef("rainy-blue")) shouldBe Right(RainyBlue)
-    ConfigConvert[Color].from(ConfigValueFactory.fromAnyRef("sunny-yellow")) shouldBe Right(SunnyYellow)
-    ConfigConvert[Color].to(RainyBlue) shouldEqual ConfigValueFactory.fromAnyRef("rainy-blue")
-    ConfigConvert[Color].to(SunnyYellow) shouldEqual ConfigValueFactory.fromAnyRef("sunny-yellow")
-  }
-
-  it should "provide customizable methods to derive readers for enumerations encoded as sealed traits" in {
-    import pureconfig.generic.semiauto._
-
-    implicit val colorReader = deriveEnumerationReader[Color](ConfigFieldMapping(PascalCase, SnakeCase))
-
-    ConfigReader[Color].from(ConfigValueFactory.fromAnyRef("rainy_blue")) shouldBe Right(RainyBlue)
-    ConfigReader[Color].from(ConfigValueFactory.fromAnyRef("sunny_yellow")) shouldBe Right(SunnyYellow)
-  }
-
-  it should "provide customizable methods to derive writers for enumerations encoded as sealed traits" in {
-    import pureconfig.generic.semiauto._
-
-    implicit val colorWriter = deriveEnumerationWriter[Color](ConfigFieldMapping(PascalCase, SnakeCase))
-
-    ConfigWriter[Color].to(RainyBlue) shouldEqual ConfigValueFactory.fromAnyRef("rainy_blue")
-    ConfigWriter[Color].to(SunnyYellow) shouldEqual ConfigValueFactory.fromAnyRef("sunny_yellow")
-  }
-
-  it should "provide customizable methods to derive full converters for enumerations encoded as sealed traits" in {
-    import pureconfig.generic.semiauto._
-
-    implicit val colorConvert = deriveEnumerationConvert[Color](ConfigFieldMapping(PascalCase, SnakeCase))
-
-    ConfigConvert[Color].from(ConfigValueFactory.fromAnyRef("rainy_blue")) shouldBe Right(RainyBlue)
-    ConfigConvert[Color].from(ConfigValueFactory.fromAnyRef("sunny_yellow")) shouldBe Right(SunnyYellow)
-    ConfigConvert[Color].to(RainyBlue) shouldEqual ConfigValueFactory.fromAnyRef("rainy_blue")
-    ConfigConvert[Color].to(SunnyYellow) shouldEqual ConfigValueFactory.fromAnyRef("sunny_yellow")
-  }
-
-  it should "not allow deriving readers, writers and full converters for enumerations encoded as sealed traits whose subclasses are not all case objects" in {
-    import pureconfig.generic.semiauto._
-
-    illTyped("deriveEnumerationReader[Entity]")
-    illTyped("deriveEnumerationWriter[Entity]")
-    illTyped("deriveEnumerationConvert[Entity]")
   }
 
   behavior of "auto"
